upload music complete

+++ src/main/java/com/TLU/SoundVerse/entity/Music.java
@@ -34,7 +34,7 @@ public class Music {
 
     private Integer genreId;
 
-    private Integer length;
+    private Float length;
 
     private String filePath;
